service-gateway/requirements.txt
anyio==3.6.2
async-timeout==4.0.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
crc16==0.1.1
cryptography==38.0.3
Deprecated==1.2.13
fastapi==0.87.0
grpcio==1.50.0
grpcio-status==1.50.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
limits==2.7.1
msgpack==1.0.4
numpy==1.23.5
opencv-python==4.6.0.66
packaging==21.3
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pylogbeat==2.0.0
pyparsing==3.0.9
python-logstash-async==2.5.0
python-redis==0.3.0
redis==4.3.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.20.0
wrapt==1.14.1
